src/llama_index/gcf/create_weaviate_schema/requirements.txt
functions-framework==3.*
weaviate-client==3.24.2